% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_stats.R
\name{export_stats}
\alias{export_stats}
\title{Ce code sert à exporter les tableaux statistiques produites à partir de l'open data / fichier ZAPM}
\usage{
export_stats(file, mode_traitement, data_export, orange_file = NULL)
}
\description{
Ce code sert à exporter les tableaux statistiques produites à partir de l'open data / fichier ZAPM
}
