% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiche_erreurs.R
\name{fiche_erreurs}
\alias{fiche_erreurs}
\title{code pour obtenir la fiche d'erreurs des jointures effectuées entre data (dernière version avicca, open data ARCEP, fichier trajectoire Orange)}
\usage{
fiche_erreurs(avicca, arcep = NULL, orange = NULL)
}
\description{
code pour obtenir la fiche d'erreurs des jointures effectuées entre data (dernière version avicca, open data ARCEP, fichier trajectoire Orange)
}
