% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointure_data.R
\name{jointure}
\alias{jointure}
\title{cette fonction programme la jointure entre les différentes sources de données pour obtenir l'open data Avicca}
\usage{
jointure(avicca, arcep = NULL, orange = NULL)
}
\description{
cette fonction programme la jointure entre les différentes sources de données pour obtenir l'open data Avicca
}
