% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_open_data_avicca.R
\name{lecture_avicca}
\alias{lecture_avicca}
\title{Ce code permet d'ouvrir dans R l'open data Avicca}
\usage{
lecture_avicca(path)
}
\description{
Ce code permet d'ouvrir dans R l'open data Avicca
}
