% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitement_fichier_ZAPM.R
\name{manip_fichier_zapm}
\alias{manip_fichier_zapm}
\title{Ce code concerne les manipulations sur le fichier ZAPM
(1) Transformer les données en format Excel avec quelques filtres selon le type de ZAPM (>=10 ans <=50\% couverture, etc)
(2) Produire quelques stats : Taille des ZAPM par département selon certains type de ZAPM; Stock de ZAPM (Stock ZAPM >= 5 ans et couverture <= 95\%  ou 80\%)}
\usage{
manip_fichier_zapm(avicca, chemin_fichier_zapm, name_fichier_zapm)
}
\description{
Ce code concerne les manipulations sur le fichier ZAPM
(1) Transformer les données en format Excel avec quelques filtres selon le type de ZAPM (>=10 ans <=50\% couverture, etc)
(2) Produire quelques stats : Taille des ZAPM par département selon certains type de ZAPM; Stock de ZAPM (Stock ZAPM >= 5 ans et couverture <= 95\%  ou 80\%)
}
