% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_arcep_data.R
\name{process_arcep}
\alias{process_arcep}
\title{Ce code permet d'ouvrir dans R les données de l'ARCEP sur le déploiement et extrait les colonnes que l'on recherche pour le nouveau trimestre
Colonnes: TX XXX (locaux raccordable), Estimation locaux, Source retenue, Base locaux sans construction, locaux en construction, \%Tx xxx (taux de couverture FttH)}
\usage{
process_arcep(file_path)
}
\description{
Ce code permet d'ouvrir dans R les données de l'ARCEP sur le déploiement et extrait les colonnes que l'on recherche pour le nouveau trimestre
Colonnes: TX XXX (locaux raccordable), Estimation locaux, Source retenue, Base locaux sans construction, locaux en construction, \%Tx xxx (taux de couverture FttH)
}
