% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production_resumes_statistiques.R
\name{prod_stats}
\alias{prod_stats}
\title{Ce code permet de produire les resumés statistiques pour faire les graphs sur flourish ainsi que certaines cartes notamment les cartes niveau départemental}
\usage{
prod_stats(avicca, departements, orange_data = NULL)
}
\description{
Ce code permet de produire les resumés statistiques pour faire les graphs sur flourish ainsi que certaines cartes notamment les cartes niveau départemental
}
